Added authorisation middleware
