+++ YelpCamp/middleware/auth.js
@@ -1,7 +1,41 @@
+const { campgroundSchema, reviewSchema } = require('../schemas.js');
+const ExpressError = require('../utils/ExpressError');
+const Campground = require('../models/campground');
+
 module.exports.isLoggedIn = (req, res, next) => {
     if (!req.isAuthenticated()) {
         req.flash('error', 'You must be logged in to do that!');
         return res.redirect('/login');
     }
     next();
+}
+
+module.exports.validateCampground = (req, res, next) => {
+    const { error } = campgroundSchema.validate(req.body);
+    if(error) {
+        const msg = error.details.map(el => el.message).join(', ');
+        throw new ExpressError(msg, 400);
+    } else {
+        next();
+    }
+}
+
+module.exports.isAuthor = async (req, res, next) => {
+	const { id } = req.params;
+	const campground = await Campground.findByIdAndUpdate(id);
+	if (!campground.author.equals(req.user._id)) {
+		req.flash('error', 'You are not authorized to make changes to this campground!');
+		return res.redirect(`/campgrounds/${id}`);
+	}
+	next();
+}
+
+module.exports.validateReview = (req, res, next) => {
+	const { error } = reviewSchema.validate(req.body);
+	if(error) {
+		const msg = error.details.map(el => el.message).join(', ');
+		throw new ExpressError(msg, 400);
+	} else {
+		next();
+	}
 }
