+++ YelpCamp/routes/campgrounds.js
@@ -1,20 +1,8 @@
 const express = require("express");
 const router = express.Router();
 const catchAsync = require('../utils/catchAsync');
-const ExpressError = require('../utils/ExpressError');
 const Campground = require('../models/campground');
-const { isLoggedIn } = require('../middleware/auth');
-const { campgroundSchema } = require('../schemas.js');
-
-const validateCampground = (req, res, next) => {
-    const { error } = campgroundSchema.validate(req.body);
-    if(error) {
-        const msg = error.details.map(el => el.message).join(', ');
-        throw new ExpressError(msg, 400);
-    } else {
-        next();
-    }
-}
+const { isLoggedIn, isAuthor, validateCampground } = require('../middleware/auth');
 
 router.get('/', catchAsync(async (req, res) => {
 	const campgrounds = await Campground.find({});
@@ -35,7 +23,6 @@ router.post('/', isLoggedIn, validateCampground, catchAsync(async (req, res, nex
 
 router.get('/:id', catchAsync(async (req, res) => {
 	const campground = await Campground.findById(req.params.id).populate('reviews').populate('author');
-	console.log(campground);
 	if (!campground) {
 		req.flash('error', 'Campground not found!');
 		return res.redirect('/campgrounds');
@@ -43,33 +30,24 @@ router.get('/:id', catchAsync(async (req, res) => {
 	res.render('campgrounds/show', { campground });
 }));
 
-router.get('/:id/edit', isLoggedIn, catchAsync(async (req, res) => {
+router.get('/:id/edit', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
 	const { id } = req.params;
 	const campground = await Campground.findById(id);
 	if (!campground) {
 		req.flash('error', 'Campground not found!');
 		return res.redirect('/campgrounds');
 	}
-	if (!campground.author.equals(req.user._id)) {
-		req.flash('error', 'You are not authorized to edit this campground!');
-		return res.redirect(`/campgrounds/${id}`);
-	}
 	res.render('campgrounds/edit', { campground });
 }));
 
-router.put('/:id', isLoggedIn, validateCampground, catchAsync(async (req, res) => {
+router.put('/:id', isLoggedIn, isAuthor, validateCampground, catchAsync(async (req, res) => {
 	const { id } = req.params;
-	const campground = await Campground.findByIdAndUpdate(id);
-	if (!campground.author.equals(req.user._id)) {
-		req.flash('error', 'You are not authorized to edit this campground!');
-		return res.redirect(`/campgrounds/${id}`);
-	}
-	const camp = await Campground.findByIdAndUpdate(id, { ...req.body.campground }, { new: true });
+	const campground = await Campground.findByIdAndUpdate(id, { ...req.body.campground }, { new: true });
 	req.flash('success', 'Campground updated successfully!');
 	res.redirect(`/campgrounds/${campground._id}`);
 }));
 
-router.delete('/:id', isLoggedIn, catchAsync(async (req, res) => {
+router.delete('/:id', isLoggedIn, isAuthor, catchAsync(async (req, res) => {
 	const { id } = req.params;
 	await Campground.findByIdAndDelete(id);
 	req.flash('success', 'Campground deleted successfully!');
