+++ YelpCamp/routes/reviews.js
@@ -2,19 +2,8 @@ const express = require('express');
 const router = express.Router( { mergeParams: true } );
 const Review = require('../models/review');
 const Campground = require('../models/campground');
-const { reviewSchema } = require('../schemas.js');
-const ExpressError = require('../utils/ExpressError');
 const catchAsync = require('../utils/catchAsync');
-
-const validateReview = (req, res, next) => {
-	const { error } = reviewSchema.validate(req.body);
-	if(error) {
-		const msg = error.details.map(el => el.message).join(', ');
-		throw new ExpressError(msg, 400);
-	} else {
-		next();
-	}
-}
+const { validateReview } = require('../middleware/auth');
 
 router.post('/', validateReview, catchAsync(async (req, res) => {
 	const campground = await Campground.findById(req.params.id);
